`migrate-to-v2 troubleshoot`: tons of refactoring, renamed from `debug`
